added client subscriptions on entities and email sendig for new actions

+++ src/main/java/com/tim23/fishnchill/action/dto/NewActionDto.java
@@ -19,9 +19,9 @@ public class NewActionDto {
 
     private LocalDateTime reservationEnd;
 
-    private BigDecimal actionPrice;
+    private Float actionPrice;
 
-    private BigDecimal actualPrice;
+    private Float actualPrice;
 
     private Integer numberOfGuests;
